created a libs folder with standard HTTP requests
